fix(api): chanded url redirect format to inclide city id, moved dashboard, data and settigns to dynamic path
