+++ app/src/app/[lng]/[city]/data/[step]/ActivityDataTab.tsx
