+++ app/src/app/[lng]/[city]/data/[step]/page.tsx
