+++ app/src/app/[lng]/[city]/data/[step]/subsectors.json
