+++ app/src/app/[lng]/[city]/data/[step]/types.d.ts
