+++ app/src/util/types.d.ts
@@ -1,7 +1,7 @@
 import type {
   DataSourceWithRelations,
   InventoryValueData,
-} from "@/app/[lng]/data/[step]/types";
+} from "@/app/[lng]/[city]/data/[step]/types";
 import type { ScopeAttributes } from "@/models/Scope";
 import type { SectorAttributes } from "@/models/Sector";
 import type { SubCategoryAttributes } from "@/models/SubCategory";
